:truck: mv: move constants
