+++ src/app/page.tsx
@@ -7,7 +7,7 @@ import Intro from '@/components/ui/Intro/Intro';
 import PageNavigation from '@/components/ui/PageNavigation/PageNavigation';
 import SocialIcons from '@/components/ui/SocialIcons/SocialIcons';
 
-import { HOME_PROFILE_TEXT } from '@/common/constants/home';
+import { HOME_PROFILE_TEXT } from '@/constants/home';
 
 import type { NextPage } from 'next';
