+++ src/components/layout/Footer/Footer.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 
-import { PROFILE_TEXT } from '@/common/constants/about';
+import { PROFILE_TEXT } from '@/constants/about';
 
 import Footer from './Footer';
