+++ src/components/ui/Intro/Intro.test.tsx
@@ -3,7 +3,7 @@ import '@/hooks/useMediaQuery/useMediaQueryMock';
 
 import useMediaQuery from '@/hooks/useMediaQuery/useMediaQuery';
 
-import { HOME_PROFILE_TEXT } from '@/common/constants/home';
+import { HOME_PROFILE_TEXT } from '@/constants/home';
 
 import Intro from './Intro';
