+++ src/constants/skillIcons.ts
